Init project from AI prompt I have an android tablet I want to use a photo frame to show all my photos, as a slide show. All my photos are stored on my NAS local network and can be access on the network. The NAS also runs PhotoPrism which has access to all my photos, it also has an api https://docs.photoprism.dev/. I have a Home Assistant runing at homeassistant.local, which has a API, and have setup light, weather, spotify, Google TV all setup.
I want a slideshow with the photos, that is GPU accelated, simple stylish transitions, scales images to fit with no crop, blurs the background behind the image, photos are selected randomly favouring the newest photos while trying to keep to a theme (album or tag like landscape), also favour images that are in the same orentation as the tablet. It needs to support swiping next and prev.
Overlayed on the slideshow please show, location of the photo and date of the photo, todays time, weather for Stafford UK (from HA) (clicking it opens the weather app), todays reminders and events (from HA), what is playing on spotify (from HA) (with next song button, and a link to open the spotify app) if anything, if something is playing on the TV (use HA, Shield). Also well as simple button switches, like lights and start vaucum cleaner.

This needs to be written as a PWA with simple offline mode that has two or three images cached.

For tech try to keep lightweight, no server should be needed, use lit.dev and vite for build tool.